requirements.txt
paho-mqtt
supabase
python-dotenv
streamlit
pandas
plotly
watchdog
flask
gunicorn
pyserial
